// applications/frc-website/src/components/flat-roof-calculator/types/FormValues.tsx
export type FormValues = {
  guarantee?: string;
  combustible?: string;
  insulated?: string;
  color?: string;
  companyName?: string;
  projectName?: string;
  fieldArea?: number;
  upstandHeight?: number;
  upstandLength?: number;
  kerbHeight?: number;
  kerbLength?: number;
  detailHeight1?: number;
  detailLength1?: number;
  detailHeight2?: number;
  detailLength2?: number;
};